refactor: moved layout to ApLayout component
